Alarms are now shown depending on current time
Alarms are now seperated depending on the current time.

Recycler view has a logic error of displaying an empty view if a drug
added to database and does not show for the current time period. Will
fix this when I wake up

If database is empty, a message prompts users to add new reminder

To do:
notifications for each item in database
fix recycler view to not show empty views

+++ app/src/main/java/com/suveraapp/onload/AddSchedule.java
@@ -39,7 +39,7 @@ public class AddSchedule extends Fragment {
         View view = inflater.inflate(R.layout.fragment_add_schedule, container, false);
 
         final Calendar calendar = Calendar.getInstance();
-        schedule = new Schedule(0, 0); //create new schedule object
+        schedule = new Schedule( 0, 0); //create new schedule object
         // find NumberPicker
         // numberPicker = (NumberPicker) view.findViewById(R.id.numberPicker);
         // numberPicker.setMaxValue(5);
@@ -72,7 +72,7 @@ public class AddSchedule extends Fragment {
                         }
 
                         //update medicine amount and time in schedule object
-                        schedule.setDosage(Long.valueOf(editText.getText().toString()));
+                        schedule.setDosage(Integer.valueOf(editText.getText().toString()));
                         schedule.setTime(calendar.getTimeInMillis());
 
                         //pass through the schedule object if a medicine amount more than 0 is selected
